Add aperture correction to obs_sdss's CalibrateTask.
Add the newly standard subtask to measure aperture correction and boolean config to control it.
Override the default field that MeasureApCorr uses to determine stars that are suitable.
By default MeasureApCorr uses "calib_psfUsed", but PSF determination sets this field
and PSF determination is rarely run on SDSS data, so change the default to "calib_psfCandidate"
(which is set by star selection, which is always run on SDSS data).

Also rename the field calib_psf_used to calib_psfUsed and calib_psf_candidate
to calib_psfCandidate to match the standard CalibrateTask. This is relevant
because one of these fields is used to measure aperture correction by default
in CalibrateTask and the other by default in SdssCalibrateTask.