Update processCcd for new ProcessCcdTask
Add SdssNullIsrTask and modify the remaining code and config to use it.
Remove SDSS-specific calibrate task, since we are not using the
filter-specific configuration and the rest can be handled
with the new standard CalibrateTask.
Add new data products to the mapper: icExp, icExpBackground,
and postISRCCD (the last of which is not written by default, since it is redundant).
Update ProcessSdssCcdTask accordingly; it is now trivial
and can be eliminated once DM-4952 is implemented.
Remove override for ProcessCoaddTask, which is now gone